fixed remaining issue with small unit tests
